chore: move string acc addr
