+++ deprecated/x/cdp/integration_test.go
@@ -8,8 +8,8 @@ import (
 	tmtime "github.com/cometbft/cometbft/types/time"
 
 	"github.com/UnUniFi/chain/app"
+	ununifitypes "github.com/UnUniFi/chain/deprecated/types"
 	cdptypes "github.com/UnUniFi/chain/deprecated/x/cdp/types"
-	ununifitypes "github.com/UnUniFi/chain/types"
 	pricefeedtypes "github.com/UnUniFi/chain/x/pricefeed/types"
 )
