+++ deprecated/x/incentive/integration_test.go
@@ -6,9 +6,9 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/UnUniFi/chain/app"
+	ununifitypes "github.com/UnUniFi/chain/deprecated/types"
 	cdptypes "github.com/UnUniFi/chain/deprecated/x/cdp/types"
 	incentivetypes "github.com/UnUniFi/chain/deprecated/x/incentive/types"
-	ununifitypes "github.com/UnUniFi/chain/types"
 	pricefeedtypes "github.com/UnUniFi/chain/x/pricefeed/types"
 )
