+++ x/derivatives/keeper/market_test.go
@@ -5,7 +5,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	ununifitypes "github.com/UnUniFi/chain/types"
+	ununifitypes "github.com/UnUniFi/chain/deprecated/types"
 	pricefeedtypes "github.com/UnUniFi/chain/x/pricefeed/types"
 )
