+++ x/derivatives/keeper/yield_test.go
@@ -9,7 +9,7 @@ import (
 
 	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
 
-	ununifitypes "github.com/UnUniFi/chain/types"
+	ununifitypes "github.com/UnUniFi/chain/deprecated/types"
 	"github.com/UnUniFi/chain/x/derivatives/types"
 	pricefeedtypes "github.com/UnUniFi/chain/x/pricefeed/types"
 )
